+++ drugex/utils/__init__.py
@@ -1,3 +1,3 @@
-from .nsgaii import *
+from .pareto import *
 from .optim import *
 from .gcmol import *