+++ testing/runner/entrypoint.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+set -e
+
+exec runuser -u  drugex-test-runner -- "$@"
+
+